CORE-4164 Updated toolbar buttons on data selections.
Updated DiskResourcePresenterImpl to initially set all toolbar buttons disabled, except for uploads, since no folders are initially selected.
Updated DiskResourcePresenterImpl to use the user's home folder for uploads if no folder is selected in the navigation panel.
Split logic of ToolbarButtonVisibilitySelectionHandler into 2 new child classes, ToolbarButtonVisibilityNavigationHandler and ToolbarButtonVisibilityGridHandler.
ToolbarButtonVisibilityNavigationHandler only updates the upload, new folder, and refresh buttons when a folder is selected in the navigation panel.
ToolbarButtonVisibilityGridHandler updates the remaining buttons only when an item is selected from the main grid panel.
Minimal updates to DiskResourcePresenterTest in order to get it to pass with these changes.
Minor DiskResourceView refactor.