CORE-4164 Updated share buttons in Data Toolbar.
Updated enabled logic for share actions in ToolbarButtonVisibilityGridHandler so that they will be disabled if all selected items are not owned by the user.
Added enabled logic for the share data link action so that it's disabled if only one folder is selected.
Added hasOwner method to DiskResourceUtil.
Added set-enabled methods to DiskResourceViewToolbar for the share and data link menu item actions.