CORE-4182 Fixes for folder selection in Presenter and View.
Updated DiskResourceViewImpl::setSelectedFolder to set the selection inside the deferred ScheduledCommand, since the selection was not taking effect otherwise.
Updated DiskResourcePresenterImpl::setSelectedFolderById to lookup and select the folder in the view first before attempting to use a tree load handler.
Updated the DiskResourcePresenterImpl::go(HasOneWidget, HasId, List) method to not call go(HasOneWidget) if calling setSelectedFolderById, since go will attempt to re-select the view's currently selected folder, which may conflict with the selection of the given folder due to the deferred ScheduledCommand.