Updates to allow Tree view to load separately from center grid.
Updated FolderRpcProxy to not load files and not mask view when loading folder children.
Updated DiskResourcePresenterImpl::onFolderSelected to mask the view and load all the folder's contents from the service (so folder contents are not cached, for now).
Updated the treeLoader in DiskResourcePresenterImpl to use a basic ChildTreeStoreBinding.
Updated DiskResourceViewImpl::refreshFolder to always reload the folder's child folder in the tree, and to refresh the center grid by calling the presenter.onFolderSelected method.
Added a KLUDGE to DiskResourceViewImpl for displaying a loading icon while tree folders load their children.