Fixes for folder tree vs. grid loading.
Added logic to DiskResourcePresenterImpl::setSelectedFolderById that will trigger loading an existing parent folder of the given folder-to-select, when the folder-to-select is not already loaded.
Removed the erroneus presenter.onFolderSelected call from DiskResourceViewImpl::refreshFolder method.
Updated various handler methods in DiskResourcesEventHandler to use presenter.setSelectedFolderById instead of view.setSelectedFolder, which will handle any necessary lazy-loading, or de-selecting before selecting in order to load the center panel.
Updated the refresh and selection logic in the DiskResourcesEventHandler::onDiskResourcesMoved method.