CORE-4328 CORE-2607 Fixed data grid name and size sorting.
Updated DiskResourceColumnModel with a Comparator for DiskResource names so that column will sort Folder names before File names.
Updated the DiskResourceSizeValueProvider to return Integer types, so that sizes are sorted by their raw byte values.
Added a DiskResourceSizeCell so that file sizes still display correctly.
Added path values to the name and size cells so that the sort icon appears correctly in the grid when sorting those columns.
Added a DiskResourceColumnModel::getNameColumn method so that the column's ValueProvider and Comparator are available to DiskResourceViewImpl so that it can properly set an initial sort order to the grid.