CORE-4422 Fix for data grid cell conflicts.
Updated DiskResourceColumnModel so that its CheckBoxSelectionModel and "name" ColumnConfig are no longer Singleton.
This was causing the main Data Window to lose its checkbox column and hyperlink functionality in its "name" column when a FileSelectDialog was opened.
Added a DiskResourceColumnModel::setCheckboxColumnHidden method and refactored DiskResourceViewImpl to use the DiskResourceColumnModel::getNameColumn method, for the sake of encapsulation.