Fixed logic in DiskResourcesEventHandler::onDiskResourcesMoved method.
Additional cases needed to be handled if the currently selected folder was moved into a destination that was under the moved folder's parent, and vice versa.
In some cases a presenter.refreshFolder call would prune the folder to be viewed with presenter.setSelectedFolderById, which would cause an extra load to trigger, which caused the lazy-loading handler to fail.
Added a DiskResourceView::removeChildren method so that the appropriate folder could be pruned without being reloaded, allowing the following presenter.setSelectedFolderById call to trigger the load after lazy-loading is properly setup.
Added treeStore model checks in the DiskResourceViewImpl refreshFolder and removeChildren methods.