CORE-4456 Enabled center folder nav. in File Selectors.
Refactored DiskResourceNameCell::setHyperlinkEnabled into a setPreviewEnabled method, and to only disable Selection events and hyperlink styling for File resources when Preview is disabled.
Renamed the disableDiskResourceHyperlink method in DiskResourceView and its Presenter and Builder to disableFilePreview.
Added a 'caller' parameter to DiskResourceNameCell that is passed to DiskResourceSelectedEvents, so that DiskResourcesEventHandler only handles events triggered from a single source.