CORE-4374 Lazy loading bug fixes.
Updated SelectFolderByIdLoadHandler to report an error if a folder in the path it's loading does not exist.
Removed the SelectFolderByIdLoadHandler#verifyFolderExists method so that this handler lazy-loads as much as it can, then selects that folder, before reporting a "does not exist" error.
This allows the user to see the parent folder of the folder that no longer exists, since the full path may not fit in the error message.
Refactored the logic to initialize the paths to lazy-load from SelectFolderByIdLoadHandler#verifyFolderExists into an initPathsToLoad method.
Refactored SelectFolderByIdLoadHandler to check if the root folders of the Data Window have loaded yet before calling the new initPathsToLoad method (previously, the verifyFolderExists method would cause enough delay when the Data Window initially opened that a root folder check wasn't needed).
Minor refactor in DiskResourcePresenterImpl by adding a validParentPath method.