CORE-4600 Refactor DiskResourcePresenterImpl refresh methods.
Refactored the DiskResourcePresenterImpl#doRefresh so that it only fires a DiskResourceRefreshEvent.
Refactored the DiskResourcesEventHandler to implement DiskResourceRefreshEventHandler, which will now call DiskResourcePresenterImpl#refreshFolder.
Refactored all other classes to call only the DiskResourcePresenterImpl#doRefresh method.
Refactored the DiskResourceRefreshEvent into the ui-commons-module for use by the caching service facade, and to remove the obsolete 'tag' parameter.