CORE-4927 More refactoring to use folder paths.
Added DiskResourceView#getFolderByPath so that lazy-loading logic in SelectFolderByIdLoadHandler and DiskResourcePresenterImpl can still find a parent root folder of a folder that hasn't been loaded yet, and so the "Empty Trash" action can properly reload the Trash folder.
Added a DiskResourceView#getParentFolder method for the DiskResourcesEventHandler, so that it doesn't have to parse a folder's path to find its parent.