CORE-4914 Prevent filtered folders from loading.
Added a Folder#isFilter check to FolderRpcProxy and FolderContentsRpcProxy so that the service is not called if the folder is a filtered folder.
Removed obsolete TreeBeforeCollapseHandlerImpl since folders will remotely load in GXT 3.0.6 before the BeforeExpandItemEvent is fired, so canceling the event will not prevent filtered folders from attempting to load.