CORE-5072 Added Advanced Sharing and Copy Link buttons.
Added download URL keys to the DataLink AutoBean interface.
Refactored the "Copy" button in the DataLinkPanel into a "Copy Link" and "Advanced Sharing" button.
The "Copy Link" button now uses the "download-url" value from DataLink models.
The "Advanced Sharing" button will open the "download-page-url" in a new window.
Refactored the DataLinkPanelCell to use the "download-page-url" for displaying links.
Refactored the DiskResourceNameCell to use the "download-url" for copying links.
Minor logic refactor in the DataLinkPanel::TreeSelectionHandler.