CORE-5087 Update to save metadata template values.
Updated DiskResourceMetadataView so that values entered into metadata template forms are saved along with user-entered metadata.
Updated DiskResourceMetadataView to expand the metadata template form when template attributes are loaded, and to mask the form before fetching the attributes from the service.
Updated the Date template form field so that en example of the valid format is set as the field's empty text.