CORE-5175 Fixes for search field refresh.
DiskResourceRefreshEvent now only references a folder to refresh.
Removed the DiskResourceViewToolbar.Presenter#refreshFolder method and the DiskResourceRefreshEventHandler implementation from DiskResourcesEventHandler, so that the refresh logic can be directly handled by the DiskResourcePresenterImpl#doRefresh method.
Added DataSearchPresenter#refreshQuery and DiskResourceSearchField#updateSearch methods so that the DiskResourcePresenterImpl can refresh search results.
Updated the DiskResourceSearchField to update it's text value based on the file query field of submitted search queries.
Added the ToolbarButtonVisibilityNavigationHandler as a SubmitDiskResourceQueryEventHandler in DiskResourcePresenterImpl, so that the Refresh button will be enabled when searches are executed from the search field.